implement configurable factory
